move to server and fireballs
