+++ Content.Server/Baseball/BaseballBatSystem.cs
@@ -1,10 +1,16 @@
+using Content.Server.Weapon.Ranged.Systems;
 using Content.Shared.Item;
+using Content.Shared.Projectiles;
+using Content.Shared.Throwing;
 using Content.Shared.Weapons.Melee;
 using Robust.Shared.Audio;
 using Robust.Shared.Player;
 using Robust.Shared.Random;
+using Content.Shared.Weapons.Ranged;
+using Content.Shared.Weapons.Ranged.Systems;
 
-namespace Content.Shared.Throwing
+
+namespace Content.Server.Baseball
 {
     /// <summary>
     /// This handles...
@@ -15,6 +21,7 @@ namespace Content.Shared.Throwing
         [Dependency] private readonly EntityLookupSystem _entityLookupSystem = default!;
         [Dependency] private readonly SharedAudioSystem _audioSystem = default!;
         [Dependency] private readonly IRobustRandom _random = default!;
+        [Dependency] private readonly GunSystem _GunSystem = default!;
 
         /// <inheritdoc/>
         public override void Initialize()
@@ -41,11 +48,24 @@ namespace Content.Shared.Throwing
             {
                 if (EntityManager.HasComponent<ItemComponent>(entity))
                 {
-                    _throwingSystem.TryThrow(entity, diff + _random.NextAngle(10f, 180f).ToVec(), 10f * _random.NextFloat(2, component.WackForceMultiplier), uid);
-                    if (EntityManager.HasComponent<ThrownItemComponent>(entity))
-                        _audioSystem.Play("/Audio/Effects/hit_kick.ogg", Filter.Pvs(args.User), entity, AudioParams.Default);
-                }
 
+                    var rand = _random.Next(1, 5);
+
+
+                    if (rand < 4)
+                    {
+
+                        _throwingSystem.TryThrow(entity, diff + _random.NextAngle(10f, 180f).ToVec(), 10f * _random.NextFloat(2, component.WackForceMultiplier), uid);
+                        if (EntityManager.HasComponent<ThrownItemComponent>(entity))
+                            _audioSystem.Play("/Audio/Effects/hit_kick.ogg", Filter.Pvs(args.User), entity, AudioParams.Default);
+                        return;
+                    }
+
+                    var fireball = Spawn("ProjectileFireball", EntityManager.GetComponent<TransformComponent>(entity).Coordinates);
+                    _GunSystem.ShootProjectile(fireball, diff, args.User);
+                    EntityManager.DeleteEntity(entity);
+
+                }
             }
         }
     }
